added new seed

+++ db/seeds.rb
@@ -8,19 +8,17 @@
 
 require "open-uri"
 require "json"
-puts "Cleaning the database"
-Dose.destroy_all
-Cocktail.destroy_all
-Ingredient.destroy_all
 
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
+
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
+
+puts "Create ingredients"
 url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
-dict = open(url).read
-
-the_cocktail_db = JSON.parse(dict)
-drinks = the_cocktail_db.values.flatten
-
-drinks.each do |name|
-  Ingredient.create(
-    name: name.values.first
-  )
+ingredients = JSON.parse(open(url).read)
+ingredients["drinks"].each do |ingredient|
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
